Update layout: keep one compilation unit per file
