+++ .scalafmt.conf
@@ -1,5 +1,2 @@
 version = 3.5.8
 runner.dialect = scala213
-
-// Docstring wrapping breaks code examples
-docstrings.wrap = false
