+++ core/src/main/scala/org/typelevel/otel4s/AttributeKey.scala
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2022 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.typelevel.otel4s
+
+/** The type of value that can be set with an implementation of this key is
+  * denoted by the type parameter.
+  *
+  * @tparam A
+  *   the type of value that can be set with the key
+  */
+sealed trait AttributeKey[A] {
+  def name: String
+  def `type`: AttributeType[A]
+}
+
+object AttributeKey {
+  private class Impl[A](val name: String, val `type`: AttributeType[A])
+      extends AttributeKey[A]
+
+  def string(name: String): AttributeKey[String] =
+    new Impl(name, AttributeType.String)
+
+  def boolean(name: String): AttributeKey[Boolean] =
+    new Impl(name, AttributeType.Boolean)
+
+  def long(name: String): AttributeKey[Long] =
+    new Impl(name, AttributeType.Long)
+
+  def double(name: String): AttributeKey[Double] =
+    new Impl(name, AttributeType.Double)
+
+  def stringList(name: String): AttributeKey[List[String]] =
+    new Impl(name, AttributeType.StringList)
+
+  def booleanList(name: String): AttributeKey[List[Boolean]] =
+    new Impl(name, AttributeType.BooleanList)
+
+  def longList(name: String): AttributeKey[List[Long]] =
+    new Impl(name, AttributeType.LongList)
+
+  def doubleList(name: String): AttributeKey[List[Double]] =
+    new Impl(name, AttributeType.DoubleList)
+}
