+++ core/src/main/scala/org/typelevel/otel4s/Otel4s.scala
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2022 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.typelevel.otel4s
+
+import org.typelevel.otel4s.metrics.MeterProvider
+
+trait Otel4s[F[_]] {
+
+  /** A registry for creating named [[org.typelevel.otel4s.metrics.Meter]].
+    */
+  def meterProvider: MeterProvider[F]
+}
