+++ core/src/main/scala/org/typelevel/otel4s/metrics/SyncInstrumentBuilder.scala
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2022 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.typelevel.otel4s.metrics
+
+trait SyncInstrumentBuilder[F[_], A] {
+  type Self <: SyncInstrumentBuilder[F, A]
+
+  /** Sets the unit of measure for this instrument.
+    *
+    * @see
+    *   <a
+    *   href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#instrument-unit">Instrument
+    *   Unit</a>
+    *
+    * @param unit
+    *   the measurement unit. Must be 63 or fewer ASCII characters.
+    */
+  def withUnit(unit: String): Self
+
+  /** Sets the description for this instrument.
+    *
+    * @see
+    *   <a
+    *   href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#instrument-description">Instrument
+    *   Description</a>
+    *
+    * @param description
+    *   the description
+    */
+  def withDescription(description: String): Self
+
+  /** Creates an instrument with the given `unit` and `description` (if any).
+    */
+  def create: F[A]
+}
